// loop-auth/src/modules/auth/dtos/user-me.dto.ts
export class UserMeDto {
  id: number;
  userId: string;
  fullName: string;
  email: string;
  phone: string;
  role: string;
  gender: string;
  profilePicture: string;
  createdAt: Date;
  isEnabled: boolean;
  isVerified: boolean;
  phoneVerified: boolean;
  emailVerified: boolean;
  dob: Date;
  nid: string;
  tin: string;
}